requirements.txt
python-decouple~=3.8
httpx
aiofiles
setuptools
twine